LayoutTests/platform/chromium-mac/fast/canvas/text-globalAlpha-expected.txt

Ensure that globalAlpha is applied correctly to text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL imgdata[3] == 127 || imgdata[3] == 128 should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
